requirement.txt
asgiref==3.8.1
branca==0.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
folium==0.20.0
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.1
packaging==25.0
pillow==11.2.1
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
xyzservices==2025.4.0
